JAMES-3138 Move CurrentQuotas to mailbox-api
